requirements-ingest.txt
beautifulsoup4
python-dotenv
psycopg[binary]
requests
